IGNITE-14221: Code review fixes.
