+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/sql/GridSqlConst.java
@@ -23,7 +23,7 @@ import org.h2.value.Value;
 import org.h2.value.ValueBoolean;
 import org.h2.value.ValueNull;
 
-import static org.apache.ignite.internal.processors.query.h2.sql.GridSqlStatement.secureOutputToLog;
+import static org.apache.ignite.internal.processors.query.QueryUtils.includeSensitive;
 
 /**
  * Constant value.
@@ -58,7 +58,7 @@ public class GridSqlConst extends GridSqlElement {
 
     /** {@inheritDoc} */
     @Override public String getSQL() {
-        if (secureOutputToLog())
+        if (!includeSensitive())
             return "?";
 
         return val.getSQL();
